fix "outnow" for 12.1

+++ render.py
@@ -37,7 +37,7 @@ optionParser.add_option('-v', '--verbose', action='store_true', dest='verbose', 
 optionParser.add_option('-l', '--lang', action='append', dest='lang', default=[], help='language to render')
 optionParser.add_option('-k', '--keep', action='store_true', dest='keep', default=False, help='keep SVG files')
 optionParser.add_option('-s', '--size', action='append', dest='sizes', default=[], help='sizes to render')
-optionParser.add_option('-d', '--days', dest='forced_days', default=None, help='force the amount of remaining days', metavar='DAYS')
+optionParser.add_option('-d', '--days', dest='forced_days', type='int', default=None, help='force the amount of remaining days', metavar='DAYS')
 (options, args) = optionParser.parse_args(sys.argv)
 
 if len(options.sizes) == 0:
@@ -66,9 +66,9 @@ def msg_sk(n):
 
 def msg_lt(n):
     if (n % 10 == 1) and (n != 11):
-        post = 'dienų'
+        post = u'dienų'
     else:
-        post = 'dienos'
+        post = u'dienos'
     return u'Pasirodys po', post
 
 avail = {
@@ -181,9 +181,11 @@ else:
 
 if options.forced_days != None:
     days = options.forced_days
+    seconds = 0
 else:
     diff = (RELEASE - datetime.datetime.now())
     days = diff.days
+    seconds = diff.seconds
     pass
 
 workdir = None
@@ -283,7 +285,9 @@ def render(lang, truelang, top1, top2, center, bottom1, bottom2, template_varian
             if options.keep:
                 svg_outfile = "%s/%s%s.%s.svg" % (outdir, PREFIX, var, size[2], lang)
                 shutil.copyfile(workfile, svg_outfile)
-                print "SVG saved as %s" % svg_outfile
+                if options.verbose:
+                    print "SVG saved as %s" % svg_outfile
+                    pass
                 pass
 
             if rc != 0:
@@ -292,7 +296,7 @@ def render(lang, truelang, top1, top2, center, bottom1, bottom2, template_varian
         pass
     pass
 
-def render_outnow(lang, top, bottom):
+def render_outnow(lang, truelang, top, bottom, template_variant=None):
     y = unicode(top).encode('ascii', 'xmlcharrefreplace')
     z = unicode(bottom).encode('ascii', 'xmlcharrefreplace')
 
@@ -300,42 +304,79 @@ def render_outnow(lang, top, bottom):
         sys.stdout.write("%s:" % lang)
         sys.stdout.flush()
         pass
+    
+    font_repl = None
+    if truelang in font_override:
+        font_repl = font_override[truelang]
+    else:
+        font_repl = default_font
+        pass
 
     for size in sizes:
-        template = "%s/%s-%dx%d-outnow.svg" % (outdir, PREFIX, size[0], size[1])
+        if not (size[2] in options.sizes):
+            continue
 
-        workfile = os.path.join(workdir, "work.svg")
-        out = open(workfile, "wb")
-        for line in fileinput.FileInput(template, mode="rb"):
-            line = unicode(line).replace(u"@TOP@", y).replace(u"@BOTTOM@", z)
-            out.write(line)
-            pass
-        out.close()
-
-        if options.verbose:
-            sys.stdout.write(" %s" % size[2])
-            sys.stdout.flush()
+        if y != None and len(y) > 0:
+            t = "-top"
+        else:
+            t = None
             pass
 
-        outfile = "%s/%s%s.%s.png" % (outdir, size[2], var, lang)
-        rc = subprocess.call(["inkscape", "-z", "--export-png=%s" % outfile, "--export-area-page", "-w", str(size[0]), "-h", str(size[1]), workfile], stdout=dev_null)
-        if options.keep:
-            svg_outfile = "%s/%s%s.%s.svg" % (outdir, PREFIX, var, size[2], lang)
-            shutil.copyfile(workfile, svg_outfile)
-            if options.verbose:
-                print "SVG saved as %s" % svg_outfile
+        for var in varlist:
+            if template_variant == None:
+                if t != None:
+                    template = "%s-%dx%d-outnow%s%s.svg" % (PREFIX, size[0], size[1], var, t)
+                if t == None or not os.path.exists(template):
+                    template = "%s-%dx%d-outnow%s.svg" % (PREFIX, size[0], size[1], var)
+            else:
+                if t != None:
+                    template = "%s-%dx%d-outnow%s%s-%s.svg" % (PREFIX, size[0], size[1], var, t, template_variant)
+                if t == None or not os.path.exists(template):
+                    template = "%s-%dx%d-outnow%s-%s.svg" % (PREFIX, size[0], size[1], var, template_variant)
                 pass
+
+            if not os.path.exists(template):
+                if options.verbose:
+                    print "skipping %s / %s / %s: template \"%s\" does not exist" % (lang, var, size[2], template)
+                    pass
+                continue
+
+            outfile = "%s/%s%s.%s.png" % (outdir, size[2], var, lang)
+
+            if options.verbose:
+                print "%s / %s / %s: %s -> %s" % (lang, var, size[2], template, outfile)
+                pass
+
+            workfile = os.path.join(workdir, "work.svg")
+            out = open(workfile, "wb")
+            for line in fileinput.FileInput(template, mode="rb"):
+                line = unicode(line).replace(u"@TOP@", y).replace(u"@BOTTOM@", z)
+                if lang in extra:
+                    for s, r in extra[lang].iteritems():
+                        line = line.replace(s, unicode(r).encode('ascii', 'xmlcharrefreplace'))
+                        pass
+                    pass
+                if font_repl != None:
+                    line = line.replace(font_to_replace, unicode(font_repl))
+                    pass
+                out.write(line)
+                pass
+            out.close()
+
+            rc = subprocess.call(["inkscape", "-z", "--export-png=%s" % outfile, "--export-area-page", "-w", str(size[0]), "-h", str(size[1]), workfile], stdout=dev_null)
+            if options.keep:
+                svg_outfile = "%s/%s%s.%s.svg" % (outdir, PREFIX, var, size[2], lang)
+                shutil.copyfile(workfile, svg_outfile)
+                if options.verbose:
+                    print "SVG saved as %s" % svg_outfile
+                    pass
+                pass
+
+            if rc != 0:
+                print >>sys.stderr, "ERROR: call to inkscape failed for %s" % workfile
             pass
-
-        if rc != 0:
-            print >>sys.stderr, "ERROR: call to inkscape failed for %s" % workfile
-        pass
-
-    if options.verbose:
-        print
         pass
     pass
-
 
 if options.verbose:
     print "days: %d" % (days)
@@ -345,9 +386,9 @@ if not os.path.exists(outdir):
     os.makedirs(outdir)
     pass
 
-if days == 0 and diff.seconds > 0:
+if days == 0 and seconds > 0:
     for lang in languages:
-        hours = ((diff.seconds / 3600) + 1)
+        hours = ((seconds / 3600) + 1)
         text = "%02d" % (hours)
         post2 = ""
 
@@ -371,16 +412,28 @@ if days == 0 and diff.seconds > 0:
 
 elif days <= 0:
     for lang in languages:
+        if not lang in languages:
+            continue
+
         if avail.has_key(lang):
             text = avail[lang]
         else:
             text = avail['en']
             pass
+
+        if lang in almost:
+            m = almost[lang]
+            truelang = lang
+        else:
+            m = almost['en']
+            truelang = 'en'
+            pass
+
         parts = text.split("\n")
         if len(parts) == 1:
-            render_outnow(lang, parts[0], "")
+            render_outnow(lang, truelang, parts[0], "")
         else:
-            render_outnow(lang, parts[0], parts[1])
+            render_outnow(lang, truelang, parts[0], parts[1])
             pass
         pass
     pass
@@ -392,6 +445,7 @@ else:
         text = str(days)
         post2 = ''
         pre0 = ''
+
         if callable(msg):
             pre, post = msg.__call__(days)
         elif len(msg) == 4:
@@ -434,7 +488,15 @@ else:
             pre = parts[1]
             pass
 
-        render(lang, lang, pre0, pre, text, post, post2)
+        if lang in almost:
+            m = almost[lang]
+            truelang = lang
+        else:
+            m = almost['en']
+            truelang = 'en'
+            pass
+
+        render(lang, truelang, pre0, pre, text, post, post2)
         pass
     pass
